airflow/requirements.txt
apache-airflow-providers-mysql
apache-airflow-providers-postgres
dbt-core~=1.7.0
dbt-postgres~=1.7.0
pandas
PyYAML>=6.0
